Add span support Split ACRCloudExtractTools.cs
